+++ src/Controllers/Private/TenantTeams.php
@@ -81,6 +81,7 @@ class TenantTeams extends PrivateApiController implements CrudControllerInterfac
      * @param array $params
      * @throws ApiServiceException
      * @throws BadRequestException
+     * @throws ForbiddenException
      */
     public function list(array $params): void
     {
@@ -88,6 +89,8 @@ class TenantTeams extends PrivateApiController implements CrudControllerInterfac
         $this->validatePath($params, [
             'tenant' => 'required|uuid'
         ]);
+
+        $this->validateInActiveTenant($this->user, $params['tenant']);
 
         $query_filter = [];